refactor: rewrite ingredient list screens
